requirements.txt
djangorestframework
psycopg2-binary
requests